:recycle: ensure delete-gift throws error if no id is provided

+++ backend/src/data/services/gift/delete-gift.ts
@@ -1,11 +1,13 @@
 import { GiftDTO } from "../../DTO";
 import { DeleteGiftRepository } from "../../contracts/gift";
 import { DeleteGift } from "../../../domain/useCases/gift";
+import { MissingParameterError } from "../../../domain/errors";
 
 export class DeleteGiftService implements DeleteGift {
-  constructor(private readonly deleteRepository: DeleteGiftRepository) {}
+  constructor(private readonly deleteRepository: DeleteGiftRepository) { }
 
-  async execute(id: string): Promise<GiftDTO> {
+  async execute(id?: string): Promise<GiftDTO> {
+    if (!id) throw new MissingParameterError('id')
     return this.deleteRepository.delete(id);
   }
 }
